devopsloft/certbot/requirements.txt
Click==7.1.2
python-dotenv==0.14.0
boto3==1.15.6